Snake_Game_Using_Turtle
A simple snake game on python using turtle module. Turtle module was oroginally developed to teach kids to code but it can be used to display graphics and make some easy games.

In this game there are three major components namely snake head, snake body and food. Whenever the snake head touches the food(less than a distance of 20- since every object is turtle has size 10*10), a segment is been added to the snake's head.

The snake die whenever it touches the wall or it self.

It also has score and a high-score for the user.